% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-annotation.R
\name{getBoneAnnotation}
\alias{getBoneAnnotation}
\title{Get bone annotations}
\usage{
getBoneAnnotation(bones = NULL, hub = NULL, fuzzy = TRUE)
}
\arguments{
\item{bones}{Character vector of bone names (NULL for all)}

\item{hub}{AnnotationHub object}

\item{fuzzy}{Logical; use fuzzy matching (default TRUE)}
}
\value{
data.frame of bone annotations
}
\description{
Get bone annotations
}
\examples{
getBoneAnnotation("Occipital")
}
