% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-kg.R
\name{kgNeighbors}
\alias{kgNeighbors}
\title{Get KG neighbors of an entity}
\usage{
kgNeighbors(entity, depth = 1, hub = NULL)
}
\arguments{
\item{entity}{Character; entity name}

\item{depth}{Integer; traversal depth (default 1)}

\item{hub}{AnnotationHub object}
}
\value{
data.frame of all triples within depth hops
}
\description{
Get KG neighbors of an entity
}
\examples{
head(kgNeighbors("Trapezius"))
}
