% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-hub.R
\name{print.PhysioAnnotationHub}
\alias{print.PhysioAnnotationHub}
\title{Print method for PhysioAnnotationHub}
\usage{
\method{print}{PhysioAnnotationHub}(x, ...)
}
\arguments{
\item{x}{A PhysioAnnotationHub object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for PhysioAnnotationHub
}
