% Generated by roxygen2: do not edit by hand
\name{getClinicalCodes}
\alias{getClinicalCodes}
\title{Get clinical codes for muscles}
\usage{
getClinicalCodes(muscles, system = c("icd10", "icf"), hub = NULL)
}
\arguments{
\item{muscles}{Character vector of muscle names}
\item{system}{Character; "icd10" or "icf" (default "icd10")}
\item{hub}{AnnotationHub object}
}
\value{
data.frame of matching clinical codes
}
\description{
Get clinical codes for muscles
}
