% Generated by roxygen2: do not edit by hand
\name{getMuscleAnnotation}
\alias{getMuscleAnnotation}
\title{Get muscle annotations}
\usage{
getMuscleAnnotation(muscles = NULL, hub = NULL, fuzzy = TRUE)
}
\arguments{
\item{muscles}{Character vector of muscle names (NULL for all)}
\item{hub}{AnnotationHub object}
\item{fuzzy}{Logical; use fuzzy matching (default TRUE)}
}
\value{
data.frame of muscle annotations
}
\description{
Get muscle annotations
}
