% Generated by roxygen2: do not edit by hand
\name{getNerveAnnotation}
\alias{getNerveAnnotation}
\title{Get nerve annotations}
\usage{
getNerveAnnotation(nerves = NULL, hub = NULL)
}
\arguments{
\item{nerves}{Character vector of nerve names (NULL for all)}
\item{hub}{AnnotationHub object}
}
\value{
data.frame of nerve annotations
}
\description{
Get nerve annotations
}
